+++ src/braket/pennylane_plugin/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You
 # may not use this file except in compliance with the License. A copy of
@@ -16,13 +16,9 @@ from braket.pennylane_plugin.braket_device import (  # noqa: F401
     BraketLocalQubitDevice,
 )
 from braket.pennylane_plugin.ops import (  # noqa: F401
-    ISWAP,
     PSWAP,
-    XX,
     XY,
     YY,
-    ZZ,
-    CPhaseShift,
     CPhaseShift00,
     CPhaseShift01,
     CPhaseShift10,