+++ test/unit_tests/test_braket_device.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You
 # may not use this file except in compliance with the License. A copy of
@@ -12,38 +12,64 @@
 # language governing permissions and limitations under the License.
 
 import json
+from typing import Any, Dict, Optional
 from unittest import mock
 from unittest.mock import Mock, PropertyMock, patch
 
+import braket.ir as ir
 import numpy as anp
 import pennylane as qml
 import pytest
 from braket.aws import AwsDevice, AwsDeviceType, AwsQuantumTask, AwsQuantumTaskBatch
-from braket.circuits import Circuit, Instruction, Observable, gates, result_types
+from braket.circuits import Circuit, Observable, result_types
 from braket.device_schema import DeviceActionType
+from braket.device_schema.jaqcd_device_action_properties import JaqcdDeviceActionProperties
+from braket.device_schema.simulators import GateModelSimulatorDeviceCapabilities
 from braket.devices import LocalSimulator
+from braket.simulator import BraketSimulator
+from braket.task_result import GateModelTaskResult
 from braket.tasks import GateModelQuantumTaskResult
 from pennylane import QuantumFunctionError, QubitDevice
 from pennylane import numpy as np
 from pennylane.tape import QuantumTape
 
-from braket.pennylane_plugin import (
-    ISWAP,
-    PSWAP,
-    XX,
-    XY,
-    YY,
-    ZZ,
-    BraketAwsQubitDevice,
-    BraketLocalQubitDevice,
-    CPhaseShift,
-    CPhaseShift00,
-    CPhaseShift01,
-    CPhaseShift10,
-)
+import braket.pennylane_plugin.braket_device
+from braket.pennylane_plugin import BraketAwsQubitDevice, BraketLocalQubitDevice
 from braket.pennylane_plugin.braket_device import BraketQubitDevice, Shots
 
 SHOTS = 10000
+
+ACTION_PROPERTIES = JaqcdDeviceActionProperties.parse_raw(
+    json.dumps(
+        {
+            "actionType": "braket.ir.jaqcd.program",
+            "version": ["1"],
+            "supportedOperations": ["rx", "ry", "h", "cy", "cnot"],
+            "supportedResultTypes": [
+                {"name": "StateVector", "observables": None, "minShots": 0, "maxShots": 0}
+            ],
+        }
+    )
+)
+
+GATE_MODEL_RESULT = GateModelTaskResult(
+    **{
+        "measurements": [[0, 0], [0, 0], [0, 0], [1, 1]],
+        "measuredQubits": [0, 1],
+        "taskMetadata": {
+            "braketSchemaHeader": {"name": "braket.task_result.task_metadata", "version": "1"},
+            "id": "task_arn",
+            "shots": 100,
+            "deviceId": "default",
+        },
+        "additionalMetadata": {
+            "action": {
+                "braketSchemaHeader": {"name": "braket.ir.jaqcd.program", "version": "1"},
+                "instructions": [{"control": 0, "target": 1, "type": "cnot"}],
+            },
+        },
+    }
+)
 
 RESULT = GateModelQuantumTaskResult.from_string(
     json.dumps(
@@ -100,87 +126,6 @@ CIRCUIT = (
 
 DEVICE_ARN = "baz"
 
-testdata = [
-    (qml.Hadamard, gates.H, [0], []),
-    (qml.PauliX, gates.X, [0], []),
-    (qml.PauliY, gates.Y, [0], []),
-    (qml.PauliZ, gates.Z, [0], []),
-    (qml.S, gates.S, [0], []),
-    (qml.T, gates.T, [0], []),
-    (qml.CNOT, gates.CNot, [0, 1], []),
-    (qml.CZ, gates.CZ, [0, 1], []),
-    (qml.PhaseShift, gates.PhaseShift, [0], [anp.pi]),
-    (qml.RX, gates.Rx, [0], [anp.pi]),
-    (qml.RY, gates.Ry, [0], [anp.pi]),
-    (qml.RZ, gates.Rz, [0], [anp.pi]),
-    (qml.SWAP, gates.Swap, [0, 1], []),
-    (qml.CSWAP, gates.CSwap, [0, 1, 2], []),
-    (qml.Toffoli, gates.CCNot, [0, 1, 2], []),
-    (qml.QubitUnitary, gates.Unitary, [0], [np.array([[0, 1], [1, 0]])]),
-    (qml.SX, gates.V, [0], []),
-    (qml.CY, gates.CY, [0, 1], []),
-    (CPhaseShift, gates.CPhaseShift, [0, 1], [anp.pi]),
-    (CPhaseShift00, gates.CPhaseShift00, [0, 1], [anp.pi]),
-    (CPhaseShift01, gates.CPhaseShift01, [0, 1], [anp.pi]),
-    (CPhaseShift10, gates.CPhaseShift10, [0, 1], [anp.pi]),
-    (ISWAP, gates.ISwap, [0, 1], []),
-    (PSWAP, gates.PSwap, [0, 1], [anp.pi]),
-    (XY, gates.XY, [0, 1], [anp.pi]),
-    (XX, gates.XX, [0, 1], [anp.pi]),
-    (YY, gates.YY, [0, 1], [anp.pi]),
-    (ZZ, gates.ZZ, [0, 1], [anp.pi]),
-]
-
-testdata_inverses = [
-    (qml.Hadamard, gates.H, [], [], [0]),
-    (qml.PauliX, gates.X, [], [], [0]),
-    (qml.PauliY, gates.Y, [], [], [0]),
-    (qml.PauliZ, gates.Z, [], [], [0]),
-    (qml.Hadamard, gates.H, [], [], [0]),
-    (qml.S, gates.Si, [], [], [0]),
-    (qml.T, gates.Ti, [], [], [0]),
-    (qml.SX, gates.Vi, [], [], [0]),
-    (qml.CNOT, gates.CNot, [], [], [0, 1]),
-    (qml.CZ, gates.CZ, [], [], [0, 1]),
-    (qml.CY, gates.CY, [], [], [0, 1]),
-    (qml.SWAP, gates.Swap, [], [], [0, 1]),
-    (qml.CSWAP, gates.CSwap, [], [], [0, 1, 2]),
-    (qml.Toffoli, gates.CCNot, [], [], [0, 1, 2]),
-    (qml.RX, gates.Rx, [0.15], [-0.15], [0]),
-    (qml.RY, gates.Ry, [0.15], [-0.15], [0]),
-    (qml.RZ, gates.Rz, [0.15], [-0.15], [0]),
-    (qml.PhaseShift, gates.PhaseShift, [0.15], [-0.15], [0]),
-    (
-        qml.QubitUnitary,
-        gates.Unitary,
-        [
-            1
-            / anp.sqrt(2)
-            * np.array(
-                [[1, 0, 0, 1j], [0, 1j, 1, 0], [0, 1j, -1, 0], [1, 0, 0, -1j]], dtype=complex
-            )
-        ],
-        [
-            1
-            / anp.sqrt(2)
-            * np.array(
-                [[1, 0, 0, 1], [0, -1j, -1j, 0], [0, 1, -1, 0], [-1j, 0, 0, 1j]], dtype=complex
-            )
-        ],
-        [0, 1],
-    ),
-    (CPhaseShift, gates.CPhaseShift, [0.15], [-0.15], [0, 1]),
-    (CPhaseShift00, gates.CPhaseShift00, [0.15], [-0.15], [0, 1]),
-    (CPhaseShift01, gates.CPhaseShift01, [0.15], [-0.15], [0, 1]),
-    (CPhaseShift10, gates.CPhaseShift10, [0.15], [-0.15], [0, 1]),
-    (ISWAP, gates.PSwap, [], [3 * anp.pi / 2], [0, 1]),
-    (PSWAP, gates.PSwap, [0.15], [-0.15], [0, 1]),
-    (XY, gates.XY, [0.15], [-0.15], [0, 1]),
-    (XX, gates.XX, [0.15], [-0.15], [0, 1]),
-    (YY, gates.YY, [0.15], [-0.15], [0, 1]),
-    (ZZ, gates.ZZ, [0.15], [-0.15], [0, 1]),
-]
-
 
 def test_reset():
     """Tests that the members of the device are cleared on reset."""
@@ -193,27 +138,15 @@ def test_reset():
     assert dev.task is None
 
 
-@pytest.mark.parametrize("pl_op, braket_gate, qubits, params", testdata)
-def test_apply(pl_op, braket_gate, qubits, params):
+def test_apply():
     """Tests that the correct Braket gate is applied for each PennyLane operation."""
-    dev = _aws_device(wires=len(qubits))
-    circuit = dev.apply([pl_op(*params, wires=qubits)])
-    assert circuit == Circuit().add_instruction(Instruction(braket_gate(*params), qubits))
-
-
-@pytest.mark.parametrize("pl_op, braket_gate, params, inv_params, qubits", testdata_inverses)
-def test_apply_inverse_gates(pl_op, braket_gate, params, inv_params, qubits):
-    """
-    Tests that the correct Braket gate is applied for the inverse of each PennyLane operations
-    where the inverse is defined.
-    """
-    dev = _aws_device(wires=len(qubits))
-    circuit = dev.apply([pl_op(*params, wires=qubits).inv()])
-    assert circuit == Circuit().add_instruction(Instruction(braket_gate(*inv_params), qubits))
+    dev = _aws_device(wires=2)
+    circuit = dev.apply([qml.Hadamard(wires=0), qml.CNOT(wires=[0, 1])])
+    assert circuit == Circuit().h(0).cnot(0, 1)
 
 
 def test_apply_unused_qubits():
-    """Tests that the correct circuit is created when not all qires in the dievce are used."""
+    """Tests that the correct circuit is created when not all wires in the device are used."""
     dev = _aws_device(wires=4)
     operations = [qml.Hadamard(wires=1), qml.CNOT(wires=[1, 2]), qml.RX(np.pi / 2, wires=2)]
     rotations = [qml.RY(np.pi, wires=1)]
@@ -475,33 +408,37 @@ def test_simulator_none_shots():
     assert dev.analytic
 
 
-def test_local_default_shots():
+@pytest.mark.parametrize("backend", ["default", "braket_sv", "braket_dm"])
+def test_local_default_shots(backend):
     """Tests that simulator devices are analytic if ``shots`` is not supplied"""
-    dev = BraketLocalQubitDevice(wires=2)
+    dev = BraketLocalQubitDevice(wires=2, backend=backend)
     assert dev.shots is None
     assert dev.analytic
 
 
-def test_local_zero_shots():
+@pytest.mark.parametrize("backend", ["default", "braket_sv", "braket_dm"])
+def test_local_zero_shots(backend):
     """Test that the local simulator device is analytic if ``shots=0``"""
-    dev = BraketLocalQubitDevice(wires=2, shots=0)
+    dev = BraketLocalQubitDevice(wires=2, backend=backend, shots=0)
     assert dev.shots is None
     assert dev.analytic
 
 
-def test_local_none_shots():
+@pytest.mark.parametrize("backend", ["default", "braket_sv", "braket_dm"])
+def test_local_none_shots(backend):
     """Tests that the simulator devices are analytic if ``shots`` is specified to be `None`."""
-    dev = BraketLocalQubitDevice(wires=2, shots=None)
+    dev = BraketLocalQubitDevice(wires=2, backend=backend, shots=None)
     assert dev.shots is None
     assert dev.analytic
 
 
 @patch.object(LocalSimulator, "run")
 @pytest.mark.parametrize("shots", [0, 1000])
-def test_local_qubit_execute(mock_run, shots):
+@pytest.mark.parametrize("backend", ["default", "braket_sv", "braket_dm"])
+def test_local_qubit_execute(mock_run, shots, backend):
     """Tests that the local qubit device is run with the correct arguments"""
     mock_run.return_value = TASK
-    dev = BraketLocalQubitDevice(wires=4, shots=shots, foo="bar")
+    dev = BraketLocalQubitDevice(wires=4, backend=backend, shots=shots, foo="bar")
 
     with QuantumTape() as circuit:
         qml.Hadamard(wires=0)
@@ -554,21 +491,142 @@ def test_wires():
             assert w == t
 
 
-@pytest.mark.xfail(raises=NotImplementedError)
-def test_run_task_unimplemented():
-    """Tests that an error is thrown when _run_task is not implemented"""
+def test_supported_ops_set(monkeypatch):
+    """Test that the supported operations are set correctly when the device is
+    created."""
+
+    test_ops = ["TestOperation"]
+    with monkeypatch.context() as m:
+        m.setattr(braket.pennylane_plugin.braket_device, "supported_operations", lambda x: test_ops)
+        dev = _aws_device(wires=2)
+        assert dev.operations == test_ops
+
+
+def test_projection():
+    """Test that the Projector observable is correctly supported."""
+    wires = 2
+    dev = BraketLocalQubitDevice(wires=wires)
+
+    thetas = [1.5, 1.6]
+    p_01 = np.cos(thetas[0] / 2) ** 2 * np.sin(thetas[1] / 2) ** 2
+    p_10 = np.sin(thetas[0] / 2) ** 2 * np.cos(thetas[1] / 2) ** 2
+
+    def f(thetas, **kwargs):
+        [qml.RY(thetas[i], wires=i) for i in range(wires)]
+
+    measure_types = ["expval", "var", "sample"]
+    projector_01 = qml.Projector([0, 1], wires=range(wires))
+    projector_10 = qml.Projector([1, 0], wires=range(wires))
+
+    # 01 case
+    fs = [qml.map(f, [projector_01], dev, measure=m) for m in measure_types]
+    assert np.allclose(fs[0](thetas), p_01)
+    assert np.allclose(fs[1](thetas), p_01 - p_01 ** 2)
+
+    samples = fs[2](thetas, shots=100)[0].tolist()
+    assert set(samples) == {0, 1}
+
+    # 10 case
+    fs = [qml.map(f, [projector_10], dev, measure=m) for m in measure_types]
+    assert np.allclose(fs[0](thetas), p_10)
+    assert np.allclose(fs[1](thetas), p_10 - p_10 ** 2)
+
+    samples = fs[2](thetas, shots=100)[0].tolist()
+    assert set(samples) == {0, 1}
+
+
+@pytest.mark.xfail(raises=AttributeError)
+def test_none_device():
+    """Tests that an error is thrown when device is not given"""
     dev = DummyLocalQubitDevice(wires=2, device=None, shots=1000)
 
     with QuantumTape() as circuit:
         qml.Hadamard(wires=0)
         qml.CNOT(wires=[0, 1])
         qml.probs(wires=[0, 1])
+    dev.execute(circuit)
 
+
+@pytest.mark.xfail(raises=NotImplementedError)
+def test_run_task_unimplemented():
+    """Tests that an error is thrown when _run_task is not implemented"""
+    dummy = DummyCircuitSimulator()
+    dev = DummyLocalQubitDevice(wires=2, device=dummy, shots=1000)
+
+    with QuantumTape() as circuit:
+        qml.Hadamard(wires=0)
+        qml.CNOT(wires=[0, 1])
+        qml.probs(wires=[0, 1])
     dev.execute(circuit)
 
 
 class DummyLocalQubitDevice(BraketQubitDevice):
     short_name = "dummy"
+
+
+class DummyCircuitSimulator(BraketSimulator):
+    def run(
+        self, program: ir.jaqcd.Program, qubits: int, shots: Optional[int], *args, **kwargs
+    ) -> Dict[str, Any]:
+        self._shots = shots
+        self._qubits = qubits
+        return GATE_MODEL_RESULT
+
+    @property
+    def properties(self) -> GateModelSimulatorDeviceCapabilities:
+        name = "braket.device_schema.simulators.gate_model_simulator_paradigm_properties"
+        input_json = {
+            "braketSchemaHeader": {
+                "name": "braket.device_schema.simulators.gate_model_simulator_device_capabilities",
+                "version": "1",
+            },
+            "service": {
+                "braketSchemaHeader": {
+                    "name": "braket.device_schema.device_service_properties",
+                    "version": "1",
+                },
+                "executionWindows": [
+                    {
+                        "executionDay": "Everyday",
+                        "windowStartHour": "09:00",
+                        "windowEndHour": "11:00",
+                    }
+                ],
+                "shotsRange": [1, 10],
+                "deviceCost": {"price": 0.25, "unit": "minute"},
+                "deviceDocumentation": {
+                    "imageUrl": "image_url",
+                    "summary": "Summary on the device",
+                    "externalDocumentationUrl": "external doc link",
+                },
+                "deviceLocation": "us-east-1",
+                "updatedAt": "2020-06-16T19:28:02.869136",
+            },
+            "action": {
+                "braket.ir.jaqcd.program": {
+                    "actionType": "braket.ir.jaqcd.program",
+                    "version": ["1"],
+                    "supportedOperations": ["x", "y", "h", "cnot"],
+                    "supportedResultTypes": [
+                        {
+                            "name": "resultType1",
+                            "observables": ["observable1"],
+                            "minShots": 2,
+                            "maxShots": 4,
+                        }
+                    ],
+                }
+            },
+            "paradigm": {
+                "braketSchemaHeader": {
+                    "name": name,
+                    "version": "1",
+                },
+                "qubitCount": 31,
+            },
+            "deviceParameters": {},
+        }
+        return GateModelSimulatorDeviceCapabilities.parse_obj(input_json)
 
 
 def _noop(*args, **kwargs):
@@ -581,7 +639,8 @@ def _noop(*args, **kwargs):
 def _aws_device(
     properties_mock, type_mock, wires, device_type=AwsDeviceType.QPU, shots=SHOTS, **kwargs
 ):
-    properties_mock.action = {DeviceActionType.JAQCD: "foo"}
+    properties_mock.action = {DeviceActionType.JAQCD: ACTION_PROPERTIES}
+    properties_mock.return_value.action.return_value = {DeviceActionType.JAQCD: ACTION_PROPERTIES}
     type_mock.return_value = device_type
     return BraketAwsQubitDevice(
         wires=wires,
@@ -596,7 +655,7 @@ def _aws_device(
 @patch.object(AwsDevice, "__init__", _noop)
 @patch.object(AwsDevice, "properties")
 def _bad_aws_device(properties_mock, wires, **kwargs):
-    properties_mock.action = {DeviceActionType.ANNEALING: "foo"}
+    properties_mock.action = {DeviceActionType.ANNEALING: ACTION_PROPERTIES}
     properties_mock.type = AwsDeviceType.QPU
     return BraketAwsQubitDevice(
         wires=wires,